requirements.txt
numpy>=1.8
numba
panda3d
pyqt
ipython
qtconsole
ipykernel
imageio
pandas
coverage
pyyaml
